requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.8.0.post1
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.8
intel-openmp==2021.4.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.42.0
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
setuptools==73.0.1
sniffio==1.3.1
sympy==1.13.2
tbb==2021.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.5
transformers==4.44.2
ttkbootstrap==1.10.1
typing_extensions==4.12.2
urllib3==2.2.2
